requirements.txt
Django==3.0
限制流量的类库
django_ratelimit==1.1.0
文件上传库
django_chunked_upload==1.1.3
图片显示工具，Django3.0暂时需要下载github上代码自己安装
sorl-thumbnail==12.5.0
图片显示
pillow==5.3.0
channels
channels_redis
celery
django-celery
redis==2.10.6
djangorestframework
代码高亮
pygments
http请求工具
httpie
# restapi 展示工具
drf-yasg
paho-mqtt
PyJWT